Added a WHOLE LOTTA STUFF :)
In this update, I have added characters and character movement, several scripts to manage the game itself, along with creating dialogue for the NPCs. Along with the NPCs, I have also implemented (albeit in a broken way) the dialogue box and the necessary tools to interact with and talk to NPCs.